+++ noodles-util/src/alignment/async.rs
@@ -1,3 +1,4 @@
 //! Async alignment format utilities.
 
+pub mod fs;
 pub mod io;